pyproject.toml
[project]
name = "kling-ai-poc"
version = "0.1.0"
description = "Kling AI POC - FastAPI Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.18",
    "pyjwt>=2.9.0",
    "httpx>=0.28.0",
]
